+++ daemon/Cargo.toml
@@ -14,3 +14,6 @@ libloading = "0.7"
 optional_take = "0.1.0"
 serde_json = "1.0"
 threadpool = "1.8"
+
+[dev-dependencies]
+encoding_rs = "0.8"
